Add files via upload
Welcome to my submission for Task 1 of the Data Science Internship at Prodigy Infotech. In this task, I have performed Exploratory Data Analysis (EDA) on a dataset provided, focusing on creating a visualization to represent the distribution of a categorical or continuous variable.

Dataset
The dataset used for this task is world_population_dataset. This dataset contains records of population from the year 2001 to 2022.

Tools and Libraries used
Jupyter notebook
Pandas
Numpy
Matplotlip & Seaborn for visualization
Exploratory Data Analysis (EDA)
During the EDA process, I performed the following steps:

Data Cleaning: Checked for missing values, duplicates, and outliers in the dataset and handled them accordingly.

Visualization: Created a bar chart, stacked chart to visualize the distribution of a categorical or continuous variable.

Conclusion
In conclusion, this EDA process provided valuable insights into the distribution of the selected variable in the dataset. This analysis lays the foundation for further exploration and modeling tasks in the data science workflow.

Thank you for reviewing my submission!

📬 Contact
For any inquiries or feedback regarding this project, please contact:

Mansi Waghmare (LinkedIn)
Email: manasiwaghmare18@gmail.com